Change architecture for giving context to the celery.
