+++ backend/web/__init__.py
@@ -1,31 +1,6 @@
-from flask import Flask
-from flask_restful import Api
-from flask_migrate import Migrate
 import os
+from celery import Celery
 
-from backend.web.api import GrechkaApi
-from backend.web.models import db, ProductItem
-
-
-def create_app():
-    """Application factory for Flask"""
-    app = Flask(__name__)
-    app.config.from_mapping(
-        SQLALCHEMY_TRACK_MODIFICATIONS=False,
-        SQLALCHEMY_DATABASE_URI=os.getenv("DB_URI"),
-        CELERY_BROKER_URL=os.getenv("CELERY_BROKER_URL"),
-        CELERY_RESULT_BACKEND=os.getenv("CELERY_RESULT_BACKEND")
-    )
-    api_instance = Api(app)
-    api_instance.add_resource(GrechkaApi, "/api/grechka")
-    db.init_app(app)
-    Migrate(app, db)
-
-    app.app_context().push()
-    db.create_all()
-
-    item = ProductItem(title="h", price=1, producer="p", prod_link="http", image_link="http")
-    db.session.add(item)
-    db.session.commit()
-
-    return app
+celery = Celery(__name__,
+                CELERY_BROKER_URL=os.getenv("CELERY_BROKER_URL"),
+                CELERY_RESULT_BACKEND=os.getenv("CELERY_RESULT_BACKEND"))
