+++ backend/web/celery_utils.py
@@ -1,0 +1,11 @@
+def init_celery(app, celery):
+    """Add flask app context to celery.Task"""
+    TaskBase = celery.Task
+
+    class ContextTask(TaskBase):
+        abstract = True
+
+        def __call__(self, *args, **kwargs):
+            with app.app_context():
+                return TaskBase.__call__(self, *args, **kwargs)
+    celery.Task = ContextTask
