+++ backend/web/uwsgi.ini
@@ -4,7 +4,7 @@ module = backend.web.wsgi_wrapper
 master = true
 processes = 5
 
-http-socket = :$(PORT)
+http-socket = 0.0.0.0:8000
 vacuum = true
 
 die-on-term = true
