+++ backend/web/wsgi_wrapper.py
@@ -1,4 +1,4 @@
-from . import create_app
+from backend.web.app import create_app
 
 application = create_app()
